refactor tests to specific classes
