+++ java/src/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/LayoutTurntableEditor.java
@@ -374,5 +374,6 @@ public class LayoutTurntableEditor extends LayoutTrackEditor {
         }
     }    
 
+
     private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutTurntableEditor.class);
 }
