+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/LayoutXOverEditorTest.java
@@ -1,9 +1,21 @@
 package jmri.jmrit.display.layoutEditor.LayoutEditorDialogs;
 
+import java.awt.Component;
+import java.awt.GraphicsEnvironment;
+import java.awt.geom.Point2D;
+import javax.swing.*;
+import jmri.*;
+import jmri.jmrit.display.EditorFrameOperator;
 import jmri.jmrit.display.layoutEditor.*;
 import jmri.util.*;
-
+import jmri.util.junit.rules.RetryRule;
+import jmri.util.swing.JemmyUtil;
 import org.junit.*;
+import org.junit.rules.Timeout;
+import org.netbeans.jemmy.ComponentChooser;
+import org.netbeans.jemmy.operators.*;
+import org.netbeans.jemmy.operators.Operator.StringComparator;
+import org.netbeans.jemmy.util.NameComponentChooser;
 
 /**
  * Test simple functioning of LayoutXOverEditor.
@@ -16,6 +28,6 @@ public class LayoutXOverEditorTest extends LayoutTurnoutEditorTest {
     public void testCtor() {
         new LayoutXOverEditor(null);
     }
-    
+
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutXOverEditorTest.class);
 }
