+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/TrackSegmentEditorTest.java
@@ -1,9 +1,21 @@
 package jmri.jmrit.display.layoutEditor.LayoutEditorDialogs;
 
+import java.awt.Component;
+import java.awt.GraphicsEnvironment;
+import java.awt.geom.Point2D;
+import javax.swing.*;
+import jmri.*;
+import jmri.jmrit.display.EditorFrameOperator;
 import jmri.jmrit.display.layoutEditor.*;
 import jmri.util.*;
-
+import jmri.util.junit.rules.RetryRule;
+import jmri.util.swing.JemmyUtil;
 import org.junit.*;
+import org.junit.rules.Timeout;
+import org.netbeans.jemmy.ComponentChooser;
+import org.netbeans.jemmy.operators.*;
+import org.netbeans.jemmy.operators.Operator.StringComparator;
+import org.netbeans.jemmy.util.NameComponentChooser;
 
 /**
  * Test simple functioning of TrackSegmentEditor.
@@ -17,5 +29,170 @@ public class TrackSegmentEditorTest extends LayoutTrackEditorTest {
         new TrackSegmentEditor(null);
     }
     
+    @Test
+    public void testEditTrackSegmentDone() {
+        Assume.assumeFalse(GraphicsEnvironment.isHeadless());
+        Assume.assumeFalse("Ignoring intermittent test", Boolean.getBoolean("jmri.skipTestsRequiringSeparateRunning"));
+
+        trackSegment.setArc(true);
+        trackSegment.setCircle(true);
+        createBlocks();
+
+        TrackSegmentEditor editor = new TrackSegmentEditor(layoutEditor);
+        
+        // Edit the track trackSegment
+        editor.editTrackSegment(trackSegment);
+        JFrameOperator jFrameOperator = new JFrameOperator(Bundle.getMessage("EditTrackSegment"));
+
+        // Select dashed
+        JLabelOperator styleLabelOperator = new JLabelOperator(jFrameOperator,
+                Bundle.getMessage("MakeLabel", Bundle.getMessage("Style")));
+        JComboBoxOperator styleComboBoxOperator = new JComboBoxOperator(
+                (JComboBox) styleLabelOperator.getLabelFor());
+        styleComboBoxOperator.selectItem(Bundle.getMessage("Dashed"));
+
+        // Select mainline
+        JComboBoxOperator mainlineComboboxOperator = new JComboBoxOperator(
+                jFrameOperator, new NameComponentChooser(Bundle.getMessage("Mainline")));
+        mainlineComboboxOperator.selectItem(Bundle.getMessage("Mainline"));
+
+        // Enable Hide
+        new JCheckBoxOperator(jFrameOperator, Bundle.getMessage("HideTrack")).doClick();
+
+        // Select block
+        JLabelOperator blockNameLabelOperator = new JLabelOperator(jFrameOperator,
+                Bundle.getMessage("BlockID"));
+        JComboBoxOperator blockComboBoxOperator = new JComboBoxOperator(
+                (JComboBox) blockNameLabelOperator.getLabelFor());
+        blockComboBoxOperator.getTextField().setText("Blk 2");
+
+        // Set arc angle
+        JLabelOperator setArcAngleLabelOperator = new JLabelOperator(
+                jFrameOperator, Bundle.getMessage("SetArcAngle"));
+        JTextFieldOperator jtxt = new JTextFieldOperator(
+                (JTextField) setArcAngleLabelOperator.getLabelFor());
+        jtxt.setText("35");
+
+        // Invoke layout block editor
+        new JButtonOperator(jFrameOperator, Bundle.getMessage("EditBlock", "")).doClick();
+
+        //TODO: frame (dialog) titles hard coded here...
+        // it should be based on Bundle.getMessage("EditBean", "Block", "DX Blk A"));
+        // but that isn't working...
+        JFrameOperator blkFO = new JFrameOperator("Edit Block Blk 2");
+        new JButtonOperator(blkFO, Bundle.getMessage("ButtonOK")).doClick();
+
+        new JButtonOperator(jFrameOperator, Bundle.getMessage("ButtonDone")).doClick();
+        jFrameOperator.waitClosed();    // make sure the dialog actually closed
+    }
+
+    @Test
+    public void testEditTrackSegmentCancel() {
+        Assume.assumeFalse(GraphicsEnvironment.isHeadless());
+        trackSegment.setArc(false);
+        trackSegment.setCircle(false);
+
+        TrackSegmentEditor editor = new TrackSegmentEditor(layoutEditor);
+        
+        // Edit the track trackSegment
+        editor.editTrackSegment(trackSegment);
+        JFrameOperator jFrameOperator = new JFrameOperator(Bundle.getMessage("EditTrackSegment"));
+
+        // Create empty block edit dialog
+        Thread segmentBlockError = JemmyUtil.createModalDialogOperatorThread(
+                Bundle.getMessage("ErrorTitle"),
+                Bundle.getMessage("ButtonOK"));  // NOI18N
+        new JButtonOperator(jFrameOperator, Bundle.getMessage("EditBlock", "")).doClick();
+        JUnitUtil.waitFor(() -> {
+            return !(segmentBlockError.isAlive());
+        }, "segmentBlockError finished");
+
+        new JButtonOperator(jFrameOperator, Bundle.getMessage("ButtonCancel")).doClick();
+        jFrameOperator.waitClosed();    // make sure the dialog actually closed
+    }
+
+    @Test
+    public void testEditTrackSegmentClose() {
+        Assume.assumeFalse(GraphicsEnvironment.isHeadless());
+
+        TrackSegmentEditor editor = new TrackSegmentEditor(layoutEditor);
+        
+        // Edit the track trackSegment
+        editor.editTrackSegment(trackSegment);
+        JFrameOperator jFrameOperator = new JFrameOperator(Bundle.getMessage("EditTrackSegment"));
+
+        new JButtonOperator(jFrameOperator, Bundle.getMessage("ButtonDone")).doClick();
+        jFrameOperator.waitClosed();    // make sure the dialog actually closed
+    }
+
+    @Test
+    public void testEditTrackSegmentError() {
+        Assume.assumeFalse(GraphicsEnvironment.isHeadless());
+        trackSegment.setArc(true);
+        trackSegment.setCircle(true);
+
+        TrackSegmentEditor editor = new TrackSegmentEditor(layoutEditor);
+        
+        // Edit the track trackSegment
+        editor.editTrackSegment(trackSegment);
+        JFrameOperator jFrameOperator = new JFrameOperator(Bundle.getMessage("EditTrackSegment"));
+
+        // Set arc angle
+        JTextFieldOperator jtxt = new JTextFieldOperator(jFrameOperator, 1);
+        jtxt.setText("abc");
+
+        new JButtonOperator(jFrameOperator, Bundle.getMessage("ButtonDone")).doClick();
+        jFrameOperator.waitClosed();    // make sure the dialog actually closed
+    }
+
+
+
+    private LayoutEditor layoutEditor = null;
+    private TrackSegment trackSegment = null;
+
+    @Before
+    public void setUp() {
+        super.setUp();
+        
+        JUnitUtil.resetProfileManager();
+        JUnitUtil.initLayoutBlockManager();
+        JUnitUtil.initInternalTurnoutManager();
+        JUnitUtil.initInternalSensorManager();
+        if (!GraphicsEnvironment.isHeadless()) {
+
+            layoutEditor = new LayoutEditor();
+            layoutEditor.setVisible(true);
+
+            Point2D point = new Point2D.Double(150.0, 100.0);
+            Point2D delta = new Point2D.Double(50.0, 10.0);
+
+            // Track Segment
+            PositionablePoint pp1 = new PositionablePoint("a", PositionablePoint.PointType.ANCHOR, point, layoutEditor);
+            point = MathUtil.add(point, delta);
+            PositionablePoint pp2 = new PositionablePoint("b", PositionablePoint.PointType.ANCHOR, point, layoutEditor);
+            trackSegment = new TrackSegment("Segment", pp1, HitPointType.POS_POINT, pp2, HitPointType.POS_POINT, false, false, layoutEditor
+            );
+        }
+    }
+
+    @After
+    public void tearDown() {
+        if (trackSegment != null) {
+            trackSegment.remove();
+            trackSegment.dispose();
+        }
+
+        if (layoutEditor != null) {
+            EditorFrameOperator efo = new EditorFrameOperator(layoutEditor);
+            efo.closeFrameWithConfirmations();
+        }
+        trackSegment = null;
+        layoutEditor = null;
+
+        JUnitUtil.resetWindows(false, false);
+        JUnitUtil.deregisterBlockManagerShutdownTask();
+        super.tearDown();
+    }
+
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(TrackSegmentEditorTest.class);
 }
